Add ToSignatureAlgorithm support for other CA implementations
